pyproject.toml
[project]
name = "nhs-data-pipeline"
version = "0.1.0"
description = "End-to-end NHS data pipeline with dbt, OMOP CDM, and Clinical NLP"
authors = [
    { name = "Greendly Guosadia", email = "greendlyguosadia@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.11,<3.14"

[tool.poetry]
packages = [{include = "src"}]

[tool.poetry.dependencies]
duckdb = "^1.5.1"
dbt-core = "^1.11.7"
dbt-duckdb = "^1.10.1"
snowflake-connector-python = "^4.4.0"
pandas = "^3.0.2"
pyarrow = "^23.0.1"
faker = "^40.12.0"
pydantic = "^2.12.5"
fastapi = "^0.135.2"
uvicorn = "^0.42.0"
python-dotenv = "^1.2.2"
prefect = "^3.6.24"
spacy = "^3.8.14"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 88
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.sqlfluff.core]
dialect = "duckdb"
templater = "dbt"
max_line_length = 120

[tool.sqlfluff.rules.capitalisation.keywords]
capitalisation_policy = "upper"

